% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quasimed.R
\name{run_quasimed}
\alias{run_quasimed}
\title{run_quasimed}
\usage{
run_quasimed(data.name, n_genes, covariate.names)
}
\arguments{
\item{data.name}{A data.frame or matrix with N x (2G + K), where N
is the number of samples, G is the number of genes (each gene contributes
two features: one for the zero component and one for the non-zero
component), and K is the number of covariates.}

\item{n_genes}{An integer value. The number of genes (G) represented
in the data.}

\item{covariate.names}{A character vector to specify the column name of covariates.}
}
\value{
A list containing: (1) number of genes; (2) estimated coefficients from
the outcome and two mediation models (M and F models in the methodology paper);
(3) standard errors corresponding to the estimated coefficients;
(4) logical vector indicating whether each gene's mediator
component (M model) is statistically significant;
(5) logical vector indicating whether each gene's zero-inflation
component (F model) is statistically significant;
(6) adjusted p-values for M and F models (joint significance test).
}
\description{
A main function for conducting causal mediation analysis
under the quasi-regression mediation framework for zero-inflated
single-cell data.
}
\examples{
\donttest{
  set.seed(123)  # Ensures reproducible cv.glmnet folds
  n_genes <- ncol(rosmap_realdat[, grep("^(M_)", colnames(rosmap_realdat))])
  run_quasimed(data.name = simulated_data, n_genes = n_genes, covariate.names = c("Z1", "Z2", "Z3"))
}
}
